backend/requirements.txt
# FastAPI Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9

# LangChain & LangGraph
langgraph>=0.2.0
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.20

# Database
pymongo>=4.6.0
motor>=3.3.0
redis>=5.0.0

# Authentication & Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
pyjwt>=2.8.0
cryptography>=42.0.0
python-dotenv>=1.0.0

# Validation & Models
pydantic>=2.7.0
pydantic-settings>=2.1.0
email-validator>=2.1.0

# PDF Generation
reportlab>=4.0.0
Pillow>=10.3.0

# HTTP & API
httpx>=0.26.0
aiohttp>=3.9.0

# Rate Limiting
slowapi>=0.1.9

# Utilities
python-dateutil>=2.8.0
faker>=22.0.0

# Testing (Phase 2)
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0

# Development
black>=24.1.0
flake8>=7.0.0
